requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.8.5
langgraph==0.2.58
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pytest==8.3.4
pytest-asyncio==0.24.0
alembic==1.13.3